+++ Sources/Controllers/ALKConversationViewController.swift
@@ -2569,8 +2569,8 @@ extension ALKConversationViewController: ALMQTTConversationDelegate {
 }
 
 extension ALKConversationViewController: ALKCustomPickerDelegate {
-    func filesSelected(images: [UIImage], videos: [String]) {
-        let fileCount = images.count + videos.count
+    func filesSelected(images: [UIImage], gifs: [String],videos: [String]) {
+        let fileCount = images.count + videos.count + gifs.count
         for index in 0 ..< fileCount {
             if index < images.count {
                 let image = images[index]
@@ -2598,8 +2598,31 @@ extension ALKConversationViewController: ALKCustomPickerDelegate {
                     return
                 }
                 viewModel.uploadImage(view: cell, indexPath: newIndexPath)
+            } else if index < gifs.count + images.count {
+                let gif = gifs[index - images.count]
+                if let size = FileManager().sizeOfFile(atPath: gif), size > ALApplozicSettings.getMaxImageSizeForUploadInMB() * 1024 {
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 1.0, execute: {
+                        self.showUploadRestrictionAlert()
+                    })
+                    continue
+                }
+
+                guard let indexPath = viewModel.sendVideo(atPath: gif, sourceType: .photoLibrary, metadata: configuration.messageMetadata).1 else { continue }
+                let newIndexPath = indexPath
+                tableView.beginUpdates()
+                tableView.insertSections(IndexSet(integer: newIndexPath.section), with: .automatic)
+                tableView.endUpdates()
+                tableView.scrollToBottom(animated: false)
+                
+                guard let cell = tableView.cellForRow(at: newIndexPath) as? ALKMyPhotoPortalCell else { return }
+                guard ALDataNetworkConnection.checkDataNetworkAvailable() else {
+                    let notificationView = ALNotificationView()
+                    notificationView.noDataConnectionNotificationView()
+                    return
+                }
+                viewModel.uploadImage(view: cell, indexPath: newIndexPath)
             } else {
-                let path = videos[index - images.count]
+                let path = videos[index - images.count - gifs.count]
                 
                 if let size = FileManager().sizeOfFile(atPath: path), size > (ALApplozicSettings.getMaxImageSizeForUploadInMB() * 1024) {
                     DispatchQueue.main.asyncAfter(deadline: .now() + 1.0, execute: {